+++ bin/stress-test/src/seeding/mod.rs
@@ -10,17 +10,20 @@ use metrics::SeedingMetrics;
 use miden_air::HashFunction;
 use miden_block_prover::LocalBlockProver;
 use miden_lib::{
-    account::{auth::RpoFalcon512, faucets::BasicFungibleFaucet, wallets::BasicWallet},
+    account::{auth::AuthRpoFalcon512, faucets::BasicFungibleFaucet, wallets::BasicWallet},
     note::create_p2id_note,
     utils::Serializable,
 };
 use miden_node_block_producer::store::StoreClient;
-use miden_node_proto::{domain::batch::BatchInputs, generated::store::rpc_client::RpcClient};
+use miden_node_proto::{domain::batch::BatchInputs, generated::rpc_store::rpc_client::RpcClient};
 use miden_node_store::{DataDirectory, GenesisState, Store};
 use miden_node_utils::tracing::grpc::OtelInterceptor;
 use miden_objects::{
-    Digest, Felt, ONE,
-    account::{Account, AccountBuilder, AccountId, AccountStorageMode, AccountType},
+    Felt, ONE, Word,
+    account::{
+        Account, AccountBuilder, AccountId, AccountStorageMode, AccountType,
+        delta::AccountUpdateDetails,
+    },
     asset::{Asset, FungibleAsset, TokenSymbol},
     batch::{BatchAccountUpdate, BatchId, ProvenBatch},
     block::{BlockHeader, BlockInputs, BlockNumber, ProposedBlock, ProvenBlock},
@@ -136,7 +139,7 @@ async fn generate_blocks(
 
     // share random coin seed and key pair for all accounts to avoid key generation overhead
     let coin_seed: [u64; 4] = rand::rng().random();
-    let rng = Arc::new(Mutex::new(RpoRandomCoin::new(coin_seed.map(Felt::new))));
+    let rng = Arc::new(Mutex::new(RpoRandomCoin::new(coin_seed.map(Felt::new).into())));
     let key_pair = {
         let mut rng = rng.lock().unwrap();
         SecretKey::with_rng(&mut *rng)
@@ -151,7 +154,7 @@ async fn generate_blocks(
         // create public accounts and notes that mint assets for these accounts
         let (pub_accounts, pub_notes) = create_accounts_and_notes(
             num_public_accounts,
-            AccountStorageMode::Private,
+            AccountStorageMode::Public,
             &key_pair,
             &rng,
             faucet.id(),
@@ -298,7 +301,7 @@ fn create_account(public_key: PublicKey, index: u64, storage_mode: AccountStorag
     let (new_account, _) = AccountBuilder::new(init_seed.try_into().unwrap())
         .account_type(AccountType::RegularAccountImmutableCode)
         .storage_mode(storage_mode)
-        .with_component(RpoFalcon512::new(public_key))
+        .with_auth_component(AuthRpoFalcon512::new(public_key))
         .with_component(BasicWallet)
         .build()
         .unwrap();
@@ -308,7 +311,7 @@ fn create_account(public_key: PublicKey, index: u64, storage_mode: AccountStorag
 /// Creates a new faucet account.
 fn create_faucet() -> Account {
     let coin_seed: [u64; 4] = rand::rng().random();
-    let mut rng = RpoRandomCoin::new(coin_seed.map(Felt::new));
+    let mut rng = RpoRandomCoin::new(coin_seed.map(Felt::new).into());
     let key_pair = SecretKey::with_rng(&mut rng);
     let init_seed = [0_u8; 32];
 
@@ -316,8 +319,8 @@ fn create_faucet() -> Account {
     let (new_faucet, _seed) = AccountBuilder::new(init_seed)
         .account_type(AccountType::FungibleFaucet)
         .storage_mode(AccountStorageMode::Private)
-        .with_component(RpoFalcon512::new(key_pair.public_key()))
         .with_component(BasicFungibleFaucet::new(token_symbol, 2, Felt::new(u64::MAX)).unwrap())
+        .with_auth_component(AuthRpoFalcon512::new(key_pair.public_key()))
         .build()
         .unwrap();
     new_faucet
@@ -367,7 +370,7 @@ fn create_consume_note_txs(
 
 /// Creates a transaction that creates an account and consumes the given input note.
 ///
-/// The account is updated with the assets from the input note, and the nonce is set to 1.
+/// The account is updated with the assets from the input note, and the nonce is incremented.
 fn create_consume_note_tx(
     block_ref: &BlockHeader,
     mut account: Account,
@@ -379,20 +382,26 @@ fn create_consume_note_tx(
         account.vault_mut().add_asset(*asset).unwrap();
     });
 
-    let (id, vault, sorage, code, _) = account.into_parts();
-    let updated_account = Account::from_parts(id, vault, sorage, code, ONE);
+    account.increment_nonce(ONE).unwrap();
+
+    let details = if account.is_public() {
+        AccountUpdateDetails::New(account.clone())
+    } else {
+        AccountUpdateDetails::Private
+    };
 
     ProvenTransactionBuilder::new(
-        updated_account.id(),
+        account.id(),
         init_hash,
-        updated_account.commitment(),
-        Digest::default(),
+        account.commitment(),
+        Word::empty(),
         block_ref.block_num(),
         block_ref.commitment(),
         u32::MAX.into(),
         ExecutionProof::new(Proof::new_dummy(), HashFunction::default()),
     )
     .add_input_notes(vec![input_note])
+    .account_update_details(details)
     .build()
     .unwrap()
 }
@@ -409,17 +418,16 @@ fn create_emit_note_tx(
     let slot = faucet.storage().get_item(2).unwrap();
     faucet
         .storage_mut()
-        .set_item(0, [slot[0], slot[1], slot[2], slot[3] + Felt::new(10)])
+        .set_item(0, [slot[0], slot[1], slot[2], slot[3] + Felt::new(10)].into())
         .unwrap();
 
-    let (id, vault, sorage, code, nonce) = faucet.clone().into_parts();
-    let updated_faucet = Account::from_parts(id, vault, sorage, code, nonce + ONE);
+    faucet.increment_nonce(ONE).unwrap();
 
     ProvenTransactionBuilder::new(
-        updated_faucet.id(),
+        faucet.id(),
         initial_account_hash,
-        updated_faucet.commitment(),
-        Digest::default(),
+        faucet.commitment(),
+        Word::empty(),
         block_ref.block_num(),
         block_ref.commitment(),
         u32::MAX.into(),