+++ crates/ntx-builder/src/store.rs
@@ -2,21 +2,15 @@ use std::time::Duration;
 
 use miden_node_proto::{
     domain::{account::NetworkAccountPrefix, note::NetworkNote},
-    errors::{ConversionError, MissingFieldHelper},
-    generated::{
-        requests::{
-            GetBlockHeaderByNumberRequest, GetCurrentBlockchainDataRequest,
-            GetNetworkAccountDetailsByPrefixRequest, GetUnconsumedNetworkNotesRequest,
-        },
-        store::ntx_builder_client as store_client,
-    },
+    errors::ConversionError,
+    generated::{self as proto, ntx_builder_store::ntx_builder_client as store_client},
     try_convert,
 };
 use miden_node_utils::tracing::grpc::OtelInterceptor;
 use miden_objects::{
     account::Account,
-    block::{BlockHeader, BlockNumber},
-    crypto::merkle::{MmrPeaks, PartialMmr},
+    block::BlockHeader,
+    crypto::merkle::{Forest, MmrPeaks, PartialMmr},
 };
 use miden_tx::utils::Deserializable;
 use thiserror::Error;
@@ -51,13 +45,16 @@ impl StoreClient {
         Self { inner: store }
     }
 
-    #[instrument(target = COMPONENT, name = "store.client.genesis_header", skip_all, err)]
-    pub async fn genesis_header_with_retry(&self) -> Result<BlockHeader, StoreError> {
+    /// Returns the block header and MMR peaks at the current chain tip.
+    #[instrument(target = COMPONENT, name = "store.client.get_latest_blockchain_data_with_retry", skip_all, err)]
+    pub async fn get_latest_blockchain_data_with_retry(
+        &self,
+    ) -> Result<Option<(BlockHeader, PartialMmr)>, StoreError> {
         let mut retry_counter = 0;
         loop {
-            match self.genesis_header().await {
+            match self.get_latest_blockchain_data().await {
                 Err(StoreError::GrpcClientError(err)) => {
-                    // exponential backoff with base 500ms and max 30s
+                    // Exponential backoff with base 500ms and max 30s.
                     let backoff = Duration::from_millis(500)
                         .saturating_mul(1 << retry_counter)
                         .min(Duration::from_secs(30));
@@ -66,7 +63,7 @@ impl StoreClient {
                         ?backoff,
                         %retry_counter,
                         %err,
-                        "store connection failed while fetching genesis header, retrying"
+                        "store connection failed while fetching latest blockchain data, retrying"
                     );
 
                     retry_counter += 1;
@@ -77,50 +74,27 @@ impl StoreClient {
         }
     }
 
-    async fn genesis_header(&self) -> Result<BlockHeader, StoreError> {
-        let response = self
-            .inner
-            .clone()
-            .get_block_header_by_number(tonic::Request::new(GetBlockHeaderByNumberRequest {
-                block_num: Some(BlockNumber::GENESIS.as_u32()),
-                include_mmr_proof: None,
-            }))
-            .await?
-            .into_inner()
-            .block_header
-            .ok_or(miden_node_proto::generated::block::BlockHeader::missing_field(
-                "block_header",
-            ))?;
-
-        BlockHeader::try_from(response).map_err(Into::into)
-    }
-
-    /// Returns the block header and MMR peaks at the chain tip if the input `block_num` is either
-    /// `None` or an outdated block number. If the input `block_num` is equal to the current
-    /// chain tip, this function returns `None`.
-    #[instrument(target = COMPONENT, name = "store.client.get_current_blockchain_data", skip_all, err)]
-    pub async fn get_current_blockchain_data(
+    #[instrument(target = COMPONENT, name = "store.client.get_latest_blockchain_data", skip_all, err)]
+    async fn get_latest_blockchain_data(
         &self,
-        block_num: Option<BlockNumber>,
     ) -> Result<Option<(BlockHeader, PartialMmr)>, StoreError> {
-        let request = tonic::Request::new(GetCurrentBlockchainDataRequest {
-            block_num: block_num.as_ref().map(BlockNumber::as_u32),
-        });
+        let request = tonic::Request::new(proto::blockchain::MaybeBlockNumber::default());
 
         let response = self.inner.clone().get_current_blockchain_data(request).await?.into_inner();
 
         match response.current_block_header {
             // There are new blocks compared to the builder's latest state
             Some(block) => {
-                let peaks = try_convert(response.current_peaks)?;
+                let peaks = try_convert(response.current_peaks).collect::<Result<_, _>>()?;
                 let header =
                     BlockHeader::try_from(block).map_err(StoreError::DeserializationError)?;
 
-                let peaks = MmrPeaks::new(header.block_num().as_usize(), peaks).map_err(|_| {
-                    StoreError::MalformedResponse(
-                        "returned peaks are not valid for the sent request".into(),
-                    )
-                })?;
+                let peaks = MmrPeaks::new(Forest::new(header.block_num().as_usize()), peaks)
+                    .map_err(|_| {
+                        StoreError::MalformedResponse(
+                            "returned peaks are not valid for the sent request".into(),
+                        )
+                    })?;
 
                 let partial_mmr = PartialMmr::from_peaks(peaks);
 
@@ -138,7 +112,10 @@ impl StoreClient {
         let mut page_token: Option<u64> = None;
 
         loop {
-            let req = GetUnconsumedNetworkNotesRequest { page_token, page_size: 128 };
+            let req = proto::ntx_builder_store::UnconsumedNetworkNotesRequest {
+                page_token,
+                page_size: 128,
+            };
             let resp = self.inner.clone().get_unconsumed_network_notes(req).await?.into_inner();
 
             let page: Vec<NetworkNote> = resp
@@ -163,7 +140,8 @@ impl StoreClient {
         &self,
         prefix: NetworkAccountPrefix,
     ) -> Result<Option<Account>, StoreError> {
-        let request = GetNetworkAccountDetailsByPrefixRequest { account_id_prefix: prefix.inner() };
+        let request =
+            proto::ntx_builder_store::AccountIdPrefix { account_id_prefix: prefix.inner() };
 
         let store_response = self
             .inner